Enabled capturing of a single photo

+++ AugendiagnoseIdea/augendiagnoseLib/src/main/java/de/jeisfeld/augendiagnoselib/activities/OrganizeNewPhotosActivity.java
@@ -287,8 +287,12 @@ public class OrganizeNewPhotosActivity extends BaseActivity {
 						@Override
 						public void onDialogPositiveClick(final DialogFragment dialog) {
 							// delete images
-							mPhotoLeft.delete();
-							mPhotoRight.delete();
+							if (mPhotoLeft != null) {
+								mPhotoLeft.delete();
+							}
+							if (mPhotoRight != null) {
+								mPhotoRight.delete();
+							}
 							finish();
 						}
